[Feat] [SCRUM-262] Setting component 디자인 #177
